refactor(bridge): extract action classifier
